refactor relevant Activities, adapters and dialogs in /data/view
