+++ app/src/androidTest/java/com/github/fribourgsdp/radio/AddPlaylistActivityTest.kt
@@ -21,6 +21,8 @@ import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.github.fribourgsdp.radio.data.Genre
 import com.github.fribourgsdp.radio.data.User
+import com.github.fribourgsdp.radio.data.view.AddPlaylistActivity
+import com.github.fribourgsdp.radio.data.view.UserProfileActivity
 import com.google.android.gms.tasks.Tasks
 import org.hamcrest.Matchers
 import org.hamcrest.Matchers.*
