+++ app/src/androidTest/java/com/github/fribourgsdp/radio/MainActivityTest.kt
@@ -1,8 +1,6 @@
 package com.github.fribourgsdp.radio
 
 import android.content.Context
-import android.content.Intent
-import androidx.test.core.app.ActivityScenario
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.espresso.Espresso
 import androidx.test.espresso.action.ViewActions
@@ -11,6 +9,7 @@ import androidx.test.espresso.intent.matcher.IntentMatchers
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.github.fribourgsdp.radio.data.view.UserProfileActivity
 import junit.framework.Assert.assertEquals
 import org.hamcrest.Matchers
 import org.junit.After
@@ -20,9 +19,6 @@ import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
-import org.mockito.Mock
-import org.mockito.Mockito
-import org.mockito.Mockito.*
 
 /**
  * Main Activity Tests
