+++ app/src/androidTest/java/com/github/fribourgsdp/radio/UserProfileActivityTest.kt
@@ -16,6 +16,9 @@ import androidx.test.espresso.matcher.ViewMatchers
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.github.fribourgsdp.radio.data.User
+import com.github.fribourgsdp.radio.data.view.MY_CLIENT_ID
+import com.github.fribourgsdp.radio.data.view.REDIRECT_URI
+import com.github.fribourgsdp.radio.data.view.UserProfileActivity
 import com.google.android.gms.tasks.Tasks
 import com.google.firebase.auth.FirebaseAuth
 import junit.framework.TestCase
