+++ app/src/androidTest/java/com/github/fribourgsdp/radio/mockimplementations/MockSongFragment.kt
@@ -1,7 +1,7 @@
 package com.github.fribourgsdp.radio.mockimplementations
 
 import com.github.fribourgsdp.radio.LyricsGetter
-import com.github.fribourgsdp.radio.SongFragment
+import com.github.fribourgsdp.radio.data.view.SongFragment
 
 class MockSongFragment : SongFragment() {
     override fun getLyricsGetter() : LyricsGetter{
