+++ app/src/main/java/com/github/fribourgsdp/radio/MainActivity.kt
@@ -10,6 +10,7 @@ import android.widget.Button
 
 import android.widget.ImageButton
 import com.github.fribourgsdp.radio.data.User
+import com.github.fribourgsdp.radio.data.view.UserProfileActivity
 
 class MainActivity : MyAppCompatActivity() {
     private val db = FirestoreDatabase()
