+++ app/src/main/java/com/github/fribourgsdp/radio/data/view/SongAdapter.kt
@@ -1,17 +1,23 @@
-package com.github.fribourgsdp.radio
+package com.github.fribourgsdp.radio.data.view
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
+import com.github.fribourgsdp.radio.OnClickListener
+import com.github.fribourgsdp.radio.R
+import com.github.fribourgsdp.radio.ViewHolder
 import com.github.fribourgsdp.radio.data.Song
 
 class SongAdapter(private val songList: List<Song>, private val listener: OnClickListener) :
     RecyclerView.Adapter<ViewHolder>(){
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
-        val itemView = LayoutInflater.from(parent.context).inflate(R.layout.song_item,
+        val itemView = LayoutInflater.from(parent.context).inflate(
+            R.layout.song_item,
             parent, false)
-        return ViewHolder(itemView, listener, itemView.findViewById(R.id.song_title_view), itemView.findViewById(R.id.song_subtitle_view))
+        return ViewHolder(itemView, listener, itemView.findViewById(R.id.song_title_view), itemView.findViewById(
+            R.id.song_subtitle_view
+        ))
     }
 
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
