+++ app/src/main/res/layout/activity_add_playlist.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AddPlaylistActivity">
+    tools:context=".data.view.AddPlaylistActivity">
 
     <EditText
         android:id="@+id/newPlaylistName"
