+++ app/src/main/res/layout/activity_playlists_fragment_holder.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PlaylistsFragmentHolderActivity">
+    tools:context=".data.view.PlaylistsFragmentHolderActivity">
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/container"
