+++ app/src/main/res/layout/activity_user_profile.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".UserProfileActivity">
+    tools:context=".data.view.UserProfileActivity">
 
     <ImageView
         android:id="@+id/userIcon"
